Keep the cart total true at any quantity
The cart page multiplied quantity by price as an int, and the cart has no
quantity cap. So a tampered quantity could overflow it. Two lines of two
billion showed a total of $8630087.68 instead of $3960000000.00, and an item
count of -294967296. Checkout was never affected, because it refuses a line
over Stripe's 999,999 before any order math runs. Nothing wrong could be
charged.

Line totals, the cart total and the item count are now long. Orders stay int,
since checkout bounds them. The money decision in PRODUCT.md now says which
side of that line each one is on. The new test in PricingTest failed before
the fix and passes after it.

A line at exactly int's maximum still wraps when it grows. That is recorded in
.shipit/open.md rather than fixed here.

59 tests behind ./verify, all green.

+++ src/main/java/dev/danvega/store/cart/Cart.java
@@ -22,9 +22,12 @@ public record Cart(@Id Long id, String cartId, Instant createdAt, Set<CartLine> 
         return lines.isEmpty();
     }
 
-    /** Total number of stickers, not number of lines. Two of one sticker counts as two. */
-    public int itemCount() {
-        return lines.stream().mapToInt(CartLine::quantity).sum();
+    /**
+     * Total number of stickers, not number of lines. Two of one sticker counts as two.
+     * Long, because nothing caps a line, so a few large lines can pass int's limit.
+     */
+    public long itemCount() {
+        return lines.stream().mapToLong(CartLine::quantity).sum();
     }
 
     public int quantityOf(Long stickerId) {
